% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surrogates.R
\name{fit_dicekriging_surrogate}
\alias{fit_dicekriging_surrogate}
\title{Fit GP using DiceKriging (aggregated observations)}
\usage{
fit_dicekriging_surrogate(
  history,
  metric,
  id_groups,
  param_names,
  covtype,
  noise,
  values,
  prev_model = NULL,
  fit_seed = NULL
)
}
\description{
Fit GP using DiceKriging (aggregated observations)
}
\keyword{internal}
