% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surrogates.R
\name{fit_hetgp_surrogate}
\alias{fit_hetgp_surrogate}
\title{Fit heteroskedastic GP using hetGP package}
\usage{
fit_hetgp_surrogate(
  history,
  metric,
  id_groups,
  param_names,
  covtype,
  prev_model = NULL,
  fit_seed = NULL
)
}
\description{
Fit heteroskedastic GP using hetGP package
}
\keyword{internal}
